backups/v5.5.0-baseline/requirements.txt
# AADS-ULoRA v5.5 Dependencies
# Core ML Framework
torch>=2.0.0
torchvision>=0.15.0

# Transformers and PEFT
transformers==4.56.0
peft>=0.8.0
accelerate>=0.20.0
bitsandbytes>=0.41.0

# Data Processing
numpy>=1.24.0
pandas>=2.0.0
Pillow>=10.0.0
opencv-python>=4.7.0
albumentations>=1.3.0

# Machine Learning
scikit-learn>=1.3.0

# Web API
fastapi>=0.100.0
uvicorn[standard]>=0.23.0
pydantic>=2.0.0
pydantic-settings>=2.0.0

# Database
sqlalchemy>=2.0.0
psycopg2-binary>=2.9.0
redis>=4.5.0

# Cloud Storage
boto3>=1.28.0
minio>=7.0.0

# Demo/UI
gradio>=3.45.0

# Utilities
tqdm>=4.65.0
matplotlib>=3.7.0
seaborn>=0.12.0
jupyter>=1.0.0
ipython>=8.0.0

# Logging and Monitoring
wandb>=0.15.0
tensorboard>=2.12.0

# Testing
pytest>=7.0.0
pytest-cov>=4.0.0

# Code Quality
black>=23.0.0
flake8>=6.0.0
mypy>=1.0.0
isort>=5.12.0

# Documentation
sphinx>=6.0.0
sphinx-rtd-theme>=1.2.0